+++ types/faqItems.ts
@@ -1,4 +1,5 @@
 export interface FAQItem {
   question: string;
   answer: JSX.Element;
+  id: string;
 }
